Replaced content of home page to content of sign-in page
