+++ src/pages/success/index.js
@@ -31,7 +31,9 @@ const SuccessPage = ({ location }) => {
   /* eslint-enable */
 
   const buttonHref =
-    language === defaultLanguage ? "/create-now" : `/${language}/create-now`;
+    language === defaultLanguage
+      ? "https://www.onlyoffice.com/docspace-registration.aspx"
+      : `https://www.onlyoffice.com/${language}/docspace-registration.aspx`;
 
   const onSubmitHandler = (e) => {
     e.preventDefault();
